requirements.txt
requests
Flask
beautifulsoup4
django
mysqlclient
pytz
djangorestframework
django-cors-headers